Add Clients routes

+++ Controllers/ControllerProperty.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using AutoMapper;
-using Backend_RentHouse_Khalifa_Sami.Data;
+using Backend_RentHouse_Khalifa_Sami.Data.PropertyData;
 using Backend_RentHouse_Khalifa_Sami.Model.Property;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
@@ -33,55 +33,57 @@ namespace Backend_RentHouse_Khalifa_Sami.Controllers
         [HttpGet("{id}", Name="GetPropertyById")]
         public ActionResult<Property> GetPropertyById(int id)
         {
-            Property commandItem = _repository.GetPropertyById(id);
-            if(commandItem != null)
-                return Ok(commandItem);
-            else
+            Property p = _repository.GetPropertyById(id);
+            if(p == null)
                 return NotFound();
+
+            return Ok(p);
         }
 
         [HttpPost]
         public ActionResult<Property> CreateProperty(Property property)
         {
+            if(property == null)
+                return NotFound();
+
             _repository.CreateProperty(property);
             return Ok(property);
             // return CreatedAtRoute(nameof(GetPropertyById), new {idProperty = property.idProperty}, property);
         }
-
+        
+        // Le patch permet de changer un seul champ au lieu de changer toute la table
         [HttpPatch("{id}")]
         public ActionResult<Property> UpdateProperty(int id, JsonPatchDocument<Property> patchDoc)
         {
-            //Erreur déja géré dans le SQLPropertyRepo
-            // Property p = _repository.GetPropertyById(id);
+
+            Property p = _repository.GetPropertyById(id);            
+            if(p==null)
+                return NotFound();
             
-            // if(p==null)
-            // {
-            //     return NotFound();
-            // }
             
             if(patchDoc != null){
-                Property p = _repository.GetPropertyById(id);
                 patchDoc.ApplyTo(p, ModelState);
 
                 if(!ModelState.IsValid)
                     return BadRequest(ModelState);
 
-                _repository.UpdateProperty(p);    
-                return Ok(p);
+            _repository.UpdateProperty(p);    
+            return Ok(p);
             }
             
-            return BadRequest();
-            
+            return BadRequest();            
         }
 
         [HttpDelete("{id}")]
         public ActionResult<Property> DeleteProperty(int id)
         {
+            Property p = _repository.GetPropertyById(id);
+            if(p == null)
+                return NotFound();
+
             _repository.DeleteProperty(id);    
             return Ok();
         }
-
-
       
     }
 }
